scripts/py/requirements.txt
baostock
pymongo
python-dateutil
tzdata
